+++ elastica/systems/protocol.py
@@ -68,15 +68,3 @@ class SymplecticSystemProtocol(SystemProtocol, SlenderBodyGeometryProtocol, Prot
     def dynamic_rates(
         self, time: np.float64, prefac: np.float64
     ) -> NDArray[np.float64]: ...
-
-
-class ExplicitSystemProtocol(SystemProtocol, SlenderBodyGeometryProtocol, Protocol):
-    # TODO: Temporarily made to handle explicit stepper.
-    # Need to be refactored as the explicit stepper is further developed.
-    def __call__(self, time: np.float64, dt: np.float64) -> np.float64: ...
-    @property
-    def state(self) -> StateType: ...
-    @state.setter
-    def state(self, state: StateType) -> None: ...
-    @property
-    def n_elems(self) -> int: ...
